move decorators, redirect lookup error to auth
