+++ backend/src/routers/pubsub.py
@@ -7,7 +7,7 @@ from src.packaging.user import main as get_data
 from src.db.functions.users import lock_user, unlock_user, update_user
 
 from src.external.pubsub.templates import publish_to_topic, parse_request
-from src.utils import fail_gracefully, pubsub_fail_gracefully
+from src.decorators import fail_gracefully, pubsub_fail_gracefully
 
 router = APIRouter()
