+++ backend/src/routers/users/db.py
@@ -6,7 +6,7 @@ from src.db.models.users import UserModel as DBUserModel
 from src.db.functions.users import login_user
 from src.db.functions.get import get_user_by_user_id
 
-from src.utils import async_fail_gracefully
+from src.decorators import async_fail_gracefully
 
 router = APIRouter()
